WebCL bug is fixed. CL code more readable.
Don't use subarrays. Both WebCL and OpenCL code has been modified in the
compute section so that it reflects a single device computation not a
multi device computation like the original software.

OpenDwarf converted the original for only one device but left a lot of
code that makes it miserable to deal with.